feat(agent): add Computer Use mode and desktop control tooling
- Add computer_use_mode agent, prompts, and computer_use_actions\n- Wire computer use / control hub tools and registry\n- Update claw/cowork/deep_research and prompt builder as needed\n- Expose mode in agents scene with i18n (en-US, zh-CN, zh-TW)\n- Document Feishu bot setup for remote connect (EN/zh-CN)
