+++ docs/remote-connect/feishu-bot-setup.zh-CN.md
@@ -1,0 +1,77 @@
+# 飞书机器人配置指南
+
+[English](./feishu-bot-setup.md)
+
+适用于通过飞书机器人完成 BitFun 远程连接配对。
+
+## 配置步骤
+
+### 第一步
+
+打开飞书开发者平台并登录：
+
+<https://open.feishu.cn/app?lang=zh-CN>
+
+### 第二步
+
+创建企业自建应用。
+
+### 第三步
+
+添加机器人能力：
+
+添加应用能力 - 机器人 - 添加
+
+### 第四步
+
+开通权限：
+
+权限管理 - 开通权限 - 搜索 `im:` - 选择所有免审权限 - 确认开通权限
+
+### 第五步
+
+复制应用凭证：
+
+凭证与基础信息 - App ID 和 App Secret
+
+### 第六步
+
+打开 BitFun 并启动飞书机器人连接：
+
+远程连接 - IM 机器人 - 飞书机器人 - 填写 App ID 和 App Secret - 连接
+
+### 第七步
+
+回到飞书开发者平台。
+
+### 第八步
+
+配置事件订阅：
+
+事件与回调 - 事件配置 - 订阅方式 - 使用长连接接收事件 - 保存
+
+然后添加消息事件：
+
+添加事件 - 搜索 `im.message` - 全选 - 确认添加
+
+### 第九步
+
+配置回调订阅：
+
+事件与回调 - 回调配置 - 订阅方式 - 使用长连接接收事件 - 保存
+
+然后添加卡片动作回调：
+
+添加回调 - 搜索 `card.action.trigger` - 选中 - 确认添加
+
+### 第十步
+
+发布机器人。
+
+### 第十一步
+
+打开飞书应用，搜索机器人名称，点击机器人打开对话框，输入任意消息并发送。
+
+### 第十二步
+
+输入 BitFun Desktop 显示的 6 位配对码，发送后等待连接成功。
