+++ src/crates/core/src/agentic/agents/registry.rs
@@ -1,6 +1,7 @@
 use super::{
-    Agent, AgenticMode, ClawMode, CodeReviewAgent, CoworkMode, DebugMode, DeepResearchAgent,
-    ExploreAgent, FileFinderAgent, GenerateDocAgent, InitAgent, PlanMode, TeamMode,
+    Agent, AgenticMode, ClawMode, CodeReviewAgent, ComputerUseMode, CoworkMode, DebugMode,
+    DeepResearchAgent, ExploreAgent, FileFinderAgent, GenerateDocAgent, InitAgent, PlanMode,
+    TeamMode,
 };
 use crate::agentic::agents::custom_subagents::{
     CustomSubagent, CustomSubagentKind, CustomSubagentLoader,
@@ -128,7 +129,8 @@ pub struct CustomSubagentDetail {
 
 fn default_model_id_for_builtin_agent(agent_type: &str) -> &'static str {
     match agent_type {
-        "agentic" | "Cowork" | "Plan" | "debug" | "Claw" | "DeepResearch" | "Team" => "auto",
+        "agentic" | "Cowork" | "ComputerUse" | "Plan" | "debug" | "Claw" | "DeepResearch"
+        | "Team" => "auto",
         _ => "primary",
     }
 }
@@ -303,6 +305,7 @@ impl AgentRegistry {
 
         // Register built-in sub-agents
         let builtin_subagents: Vec<Arc<dyn Agent>> = vec![
+            Arc::new(ComputerUseMode::new()),
             Arc::new(ExploreAgent::new()),
             Arc::new(FileFinderAgent::new()),
         ];
@@ -1068,7 +1071,7 @@ pub fn get_agent_registry() -> Arc<AgentRegistry> {
 
 #[cfg(test)]
 mod tests {
-    use super::{default_model_id_for_builtin_agent, merge_dynamic_mcp_tools};
+    use super::{default_model_id_for_builtin_agent, merge_dynamic_mcp_tools, AgentRegistry};
 
     #[test]
     fn top_level_modes_default_to_auto() {
@@ -1083,6 +1086,28 @@ mod tests {
         ] {
             assert_eq!(default_model_id_for_builtin_agent(agent_type), "auto");
         }
+    }
+
+    #[tokio::test]
+    async fn computer_use_is_builtin_subagent_not_mode() {
+        let registry = AgentRegistry::new();
+        let modes = registry.get_modes_info().await;
+        assert!(
+            !modes.iter().any(|agent| agent.id == "ComputerUse"),
+            "ComputerUse should be delegated through Task as a built-in sub-agent, not exposed as a top-level mode"
+        );
+
+        let subagents = registry.get_subagents_info(None).await;
+        let computer_use = subagents
+            .iter()
+            .find(|agent| agent.id == "ComputerUse")
+            .expect("ComputerUse should be registered as a built-in sub-agent");
+        assert!(computer_use
+            .default_tools
+            .contains(&"ControlHub".to_string()));
+        assert!(computer_use
+            .default_tools
+            .contains(&"ComputerUse".to_string()));
     }
 
     #[test]
