+++ src/web-ui/src/app/scenes/agents/agentVisibility.ts
@@ -2,7 +2,7 @@
 export const HIDDEN_AGENT_IDS = new Set<string>(['Claw']);
 
 /** Core mode agents shown in the top zone only; excluded from overview zone list and counts. */
-export const CORE_AGENT_IDS = new Set<string>(['agentic', 'Cowork']);
+export const CORE_AGENT_IDS = new Set<string>(['agentic', 'Cowork', 'ComputerUse']);
 
 /** Agents that appear in the bottom overview grid (same pool as filter chip counts). */
 export function isAgentInOverviewZone(agent: { id: string }): boolean {
