+++ src/web-ui/src/locales/zh-TW/flow-chat.json
@@ -245,6 +245,7 @@
       "Plan": "先規劃後執行，先明確需求並制定實施計劃，再進行編碼",
       "debug": "證據驅動的系統化調試：提出假設、收集運行時證據、精準定位並修復問題",
       "Cowork": "協作模式：先澄清再推進，輕量跟蹤進度，隨時驗證結果",
+      "ComputerUse": "獨立電腦操作智能體：感知本機環境，操作應用、瀏覽器與系統介面",
       "DeepResearch": "深度研究：並行派發多個子 Agent 同時調研不同章節，快速生成高質量研究報告",
       "Team": "虛擬工程團隊：CEO、工程經理、設計師、QA 負責人、安全官、發佈工程師 — 按完整衝刺流程協同工作"
     },
@@ -254,6 +255,7 @@
       "Plan": "Plan",
       "debug": "Debug",
       "Cowork": "Cowork",
+      "ComputerUse": "Computer Use",
       "DeepResearch": "Deep Research",
       "Team": "Team"
     },
